fix: resolve ruff lint errors in proxy readiness preflight
RUF046: ceil() already returns int in Python 3; drop redundant int() wrapper.
D103: add missing one-line docstrings to two public test functions.

+++ scripts/research/check_predictive_checkpoint_proxy_readiness.py
@@ -74,7 +74,7 @@ def _coerce_positive_int(value: Any, *, field: str) -> tuple[int | None, list[st
     if isinstance(value, float):
         if value < 0:
             return None, [f"{field} must be >= 0"]
-        return int(ceil(value)), []
+        return ceil(value), []
     return None, [f"{field} must be integer-like (int/float), got {type(value).__name__}"]
